Fix for 2358:	Implement M3UA stack
git-svn-id: https://mobicents.googlecode.com/svn/trunk/protocols/ss7@16914 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ m3ua/impl/src/main/java/org/mobicents/protocols/ss7/m3ua/impl/parameter/ParameterFactoryImpl.java
@@ -116,7 +116,7 @@ public class ParameterFactoryImpl implements ParameterFactory {
         return new ServiceIndicatorsImpl(inds);
     }
 
-    public TrafficModeType createTrafficModeType(long mode) {
+    public TrafficModeType createTrafficModeType(int mode) {
         return new TrafficModeTypeImpl(mode);
     }
